All the nurse checks are complete and seem to be working fine.
Started adding in the day checks, currently only have the "SRN nurse on each shift on each day" check done, but that seems to be working.

Also changed around the way the first empty shift is found to work by day rather than by nurse as this seemed to make it much faster for backtracking and changing things.

Added in a little code to track the total time taken to run so that we can keep an eye on it as well.